# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation Api Gutter Runner",
        "description": "Implémentation de Swagger",
        "contact": {
            "email": "admin@app.gutterrunner.ca"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Serveur d'API de démonstration"
        }
    ],
    "paths": {
...
        "/api/users/assignable": {
            "get": {
                "tags": [
                    "Demandes"
                ],
                "summary": "Liste des usagers à qui on peut assigner une demande",
                "description": "Liste des usagers à qui on peut assigner une demande",
                "operationId": "assignableUsersList",
                "responses": {
                    "200": {
                        "description": "Liste obtenue",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "ID",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "description": "Nom de l'usager",
                                                        "type": "string",
                                                        "example": "Catherine Henry"
                                                    },
                                                    "email": {
                                                        "description": "Email de l'usager",
                                                        "type": "string",
                                                        "example": "catherine.henry@gmail.com"
                                                    },
                                                    "roles": {
                                                        "description": "Role de l'usager",
                                                        "properties": {
                                                            "data": {
                                                                "type": "array",
                                                                "items": {
                                                                    "properties": {
                                                                        "id": {
                                                                            "description": "ID",
                                                                            "type": "string",
                                                                            "example": "995624"
                                                                        },
                                                                        "name": {
                                                                            "description": "Nom du role",
                                                                            "type": "string",
                                                                            "example": "Administrateur"
                                                                        },
                                                                        "permissions": {
                                                                            "description": "Les permissions du role",
                                                                            "properties": {
                                                                                "data": {
                                                                                    "type": "array",
                                                                                    "items": {
                                                                                        "properties": {
                                                                                            "id": {
                                                                                                "description": "ID",
                                                                                                "type": "string",
                                                                                                "example": "571433"
                                                                                            },
                                                                                            "name": {
                                                                                                "description": "Nom de la permission",
                                                                                                "type": "string",
                                                                                                "example": "Lecture seule"
                                                                                            }
                                                                                        },
                                                                                        "type": "object"
                                                                                    }
                                                                                }
                                                                            },
                                                                            "type": "object"
                                                                        }
                                                                    },
                                                                    "type": "object"
                                                                }
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Utilisateur non connecté",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "L'utilisateur n'est même pas connecté",
                                            "type": "string",
                                            "example": "Vous ne vous êtes pas encore authentifié."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}